1. 优化tracker流程，修复大量bug 2. tracker url由announce映射为tracker 3. 增加全局配置文件

+++ common/src/main/java/com/virus/pt/common/util/TrackerResponseUtils.java
@@ -13,7 +13,7 @@ import java.util.*;
  * @about <link href='http://zzyitj.xyz/'/>
  */
 public class TrackerResponseUtils {
-    public static String success(List<Peer> peers, int complete, int incomplete) {
+    public static String success(List<Peer> peers, long complete, long incomplete) {
         TrackerResponse trackerResponse = new TrackerResponse();
         trackerResponse.setComplete(complete);
         trackerResponse.setIncomplete(incomplete);
